Extract AllocationService from allocation controller
Move the allocation query, allocation-run, check-in, save/commit and
compensation logic out of AllocationController into a new
AllocationService, alongside the existing ScheduleService and
PlayerService. Bodies move verbatim; the handlers keep only routing,
gates, sqid parsing and response construction.

TierAsDict and AllocationPartyMetadata move into the service (it
produces them); PutEventManageAllocationSession moves to _common.py to
keep the controller and service from importing each other.

The duplicated party-leader subquery block, which carried a "TODO:
Extract common functionality", is now the _party_leader_subquery helper
parameterised on the time slot id - both call sites compile to identical
SQL.
