+++ convergence_games/apps/frontend/admin/services/__init__.py
@@ -1,9 +1,19 @@
+from ._allocation_service import (
+    AllocationPartyMetadata,
+    AllocationService,
+    TierAsDict,
+    provide_allocation_service,
+)
 from ._player_service import PlayerService, provide_player_service
 from ._schedule_service import ScheduleService, provide_schedule_service
 
 __all__ = [
+    "AllocationPartyMetadata",
+    "AllocationService",
     "PlayerService",
     "ScheduleService",
+    "TierAsDict",
+    "provide_allocation_service",
     "provide_player_service",
     "provide_schedule_service",
 ]
